+++ CHANGELOG.md
@@ -9,11 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- added mssql test server with seeding for kvis, ifsg and grippeweb
+
 ### Changes
 
 ### Deprecated
 
 ### Removed
+
+- BREAKING: remove support for python 3.11, 3.12, and 3.13
 
 ### Fixed