+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - new template https://github.com/robert-koch-institut/mex-template/releases/tag/2.0.0
+
 ### Deprecated
 
 ### Removed
@@ -29,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - new template https://github.com/robert-koch-institut/mex-template/releases/tag/1.5.0
+
 ### Removed
 
 - BREAKING: remove support for python 3.11, 3.12, and 3.13
@@ -50,5 +52,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - new template https://github.com/robert-koch-institut/mex-template/releases/tag/1.3.0
 - updated template to https://github.com/robert-koch-institut/mex-template/commit/944944
-
-### Security